// shared/enums.ts
/**
 * @license
 * SPDX-License-Identifier: Apache-2.0
 */

export enum UserRole {
  ADMIN = "ADMIN",
  CLIENT = "CLIENT",
  CONSULTANT = "CONSULTANT"
}

export enum ServiceCategory {
  COMPANY_REGISTRATION = "COMPANY_REGISTRATION",
  GST_SERVICES = "GST_SERVICES",
  TRADEMARK_SERVICES = "TRADEMARK_SERVICES",
  LICENSING = "LICENSING",
  TAXATION = "TAXATION",
  COMPLIANCE = "COMPLIANCE",
  BUSINESS_CONSULTANCY = "BUSINESS_CONSULTANCY"
}

export enum LeadStatus {
  NEW = "NEW",
  CONTACTED = "CONTACTED",
  QUALIFIED = "QUALIFIED",
  PROPOSAL_SENT = "PROPOSAL_SENT",
  WON = "WON",
  LOST = "LOST"
}

export enum HttpStatus {
  OK = 200,
  CREATED = 201,
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  INTERNAL_SERVER_ERROR = 500
}